mito_codec::row_converter

Module sparse

Structs

FieldWithId

Field with column id.

SparseEncoder

A special encoder for memtable.

SparsePrimaryKeyCodec

A codec for sparse key of metrics. It requires the input primary key columns are sorted by the column name in lexicographical order. It encodes the column id of the physical region.

SparsePrimaryKeyCodecInner 🔒

SparseValues

Sparse values representation.

Constants

COLUMN_ID_ENCODE_SIZE

The size of the column id in the encoded sparse row.

RESERVED_COLUMN_ID_TABLE_ID

The column id of the table id.

RESERVED_COLUMN_ID_TSID

The column id of the tsid.