mito_codec::row_converter

Trait PrimaryKeyCodecExt

pub trait PrimaryKeyCodecExt {
    // Required method
    fn encode_to_vec<'a, I>(&self, row: I, buffer: &mut Vec<u8>) -> Result<()>
       where I: Iterator<Item = ValueRef<'a>>;

    // Provided method
    fn encode<'a, I>(&self, row: I) -> Result<Vec<u8>>
       where I: Iterator<Item = ValueRef<'a>> { ... }
}

Row value encoder/decoder.

Required Methods

fn encode_to_vec<'a, I>(&self, row: I, buffer: &mut Vec<u8>) -> Result<()>
where I: Iterator<Item = ValueRef<'a>>,

Encodes rows to specific vec.

Note

Ensure the length of row iterator matches the length of fields.

Provided Methods

fn encode<'a, I>(&self, row: I) -> Result<Vec<u8>>
where I: Iterator<Item = ValueRef<'a>>,

Encodes rows to bytes.

Note

Ensure the length of row iterator matches the length of fields.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.